+++ doc/release-notes.md
@@ -5,3 +5,4 @@ Bitcoin ABC version 0.25.6 is now available from:
   <https://download.bitcoinabc.org/0.25.6/>
 
 This release includes the following features and fixes:
+ - Add a checkpoint after the May 15th, 2022 eCash upgrade